Move static source compilation into its own crate
* This is done so all of the build dependencies don't have to be fetched
  when not statically compiling libde265 with the Rust crate
* I also turned everything into a Cargo workspace so the target
  directory can be shared between the crates, decreasing build time
