+++ Cargo.toml
@@ -1,3 +1,6 @@
+[workspace]
+members = ["./libde265-sys", "./libde265-static", "libde265-static"]
+
 [package]
 name = "libde265"
 version = "0.1.0"
@@ -11,13 +14,13 @@ repository = "https://github.com/andreytkachenko/libde265-rust"
 license = "MIT OR Apache-2.0"
 
 [features]
-default = ["static"]
+default = ["system"]
 encoder = ["libde265-sys/encoder"]
 static = ["libde265-sys/static"]
 system = ["libde265-sys/system"]
 
 [dependencies]
-libde265-sys = "0.1.1"
+libde265-sys = { path = "libde265-sys", version = "0.1.1" }
 log = "0.4.20"
 thiserror = "1.0.56"
