+++ libde265-sys/Cargo.toml
@@ -13,19 +13,10 @@ build = "build.rs"
 
 [features]
 default = ["system"]
-static = []
 system = []
-generate-bindings = []
-encoder = []
+static = ["dep:libde265-static"]
+generate-bindings = ["dep:libde265-static", "libde265-static/generate-bindings"]
+encoder = ["dep:libde265-static", "libde265-static/encoder"]
 
 [dependencies]
-
-[build-dependencies]
-anyhow = "1.0.79"
-curl = "0.4.44"
-bindgen = "0.69.2"
-cc = "1.0.83"
-walkdir = "2.4.0"
-const_format = "0.2.32"
-flate2 = "1.0.28"
-tar = "0.4.40"
+libde265-static = { path = "../libde265-static", version = "0.1.0", optional = true }
